eliminar libros/imagenes no relacionados a ordenes/libros

+++ src/main/java/com/uade/tpo/demo/service/ImagenServiceImpl.java
@@ -2,7 +2,6 @@ package com.uade.tpo.demo.service;
 
 import java.util.List;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
@@ -10,14 +9,16 @@ import org.springframework.web.server.ResponseStatusException;
 import com.uade.tpo.demo.entity.Imagen;
 
 import com.uade.tpo.demo.repository.ImagenRepository;
+import com.uade.tpo.demo.repository.LibroRepository;
 
-
+import lombok.RequiredArgsConstructor;
 
 @Service
+@RequiredArgsConstructor
 public class ImagenServiceImpl implements ImagenService {
 
-    @Autowired
-    private ImagenRepository imageRepository;
+    private final ImagenRepository imageRepository;
+    private final LibroRepository libroRepository;
 
     @Override
     public Imagen create(Imagen image) {
@@ -46,9 +47,14 @@ public class ImagenServiceImpl implements ImagenService {
     @Override
     public void delete(long id) {
         Imagen imagen = imageRepository.findById(id)
-            .orElseThrow(() -> new org.springframework.web.server.ResponseStatusException(
-                org.springframework.http.HttpStatus.NOT_FOUND, "Imagen con id " + id + " no encontrada"));
-                
+            .orElseThrow(() -> new ResponseStatusException(
+                HttpStatus.NOT_FOUND, "Imagen con id " + id + " no encontrada"));
+
+        if (libroRepository.existsByImagen(imagen)) {
+            throw new ResponseStatusException(
+                HttpStatus.CONFLICT, "No se puede eliminar la imagen porque está asociada a un libro");
+        }
+
         imageRepository.delete(imagen);
     }
 }
